database schema revamp with improved constraints

+++ src/main/scala/muse/domain/mutate/DeleteComment.scala
@@ -2,4 +2,4 @@ package muse.domain.mutate
 
 import java.util.UUID
 
-final case class DeleteComment(reviewId: UUID, commentId: Int)
+final case class DeleteComment(reviewId: UUID, commentId: Long)
